// frontend/wigo-travel/src/types/theme-types.ts
export type ThemeBreakpoint = "xs" | "sm" | "md" | "lg" | "xl"
export type ThemeColor = "primary" | "secondary" | "neutral" | "disabled"
export type ThemeColorVariant = "main" | "light" | "dark"
export type FontWeight = "normal" | "bold"
export type FontColor = "white" | "primary" | "secondary" | "inherit"
export type TextVariant =
  | "h1"
  | "h2"
  | "h3"
  | "h4"
  | "h5"
  | "h6"
  | "subtitle1"
  | "subtitle2"
  | "body1"
  | "body2"
  | "caption"
  | "button"
  | "overline"
  | "inherit"